github codes/new python project/Tiago-Freitas---ByBit-BOT/requirements.txt
ccxt
python-dotenv
fastapi
uvicorn[standard]